Extract confirmation hook into generic module
